Ensure the data accumulator is 64 bits long
Newer Vivint sensors transmit messages with 64 bits of data, so the
accumulator must fit at least 64 bits.

+++ src/messaging/CRC16.h
@@ -7,14 +7,14 @@ public:
 	void setPoly(const char16_t& polynomial) {this->polynomial = polynomial;};
 	void setFinalXOR(const char16_t& fxor) {this->fxor = fxor;};
 	void push(const unsigned long int& data, const unsigned int& num_bits);
-	unsigned long int getData() const {return data_accum;};
+	unsigned long long int getData() const {return data_accum;};
 	char16_t getCRC();	// Explicitly limit the output CRC to 16 bits
 	void reset() {calc_crc = 0; data_accum = 0;};
 private:
 	void push(const unsigned long int& data, const unsigned int& num_bits, const bool& accumulate);
 	char16_t polynomial {0};
 	char16_t fxor {0};	// Final XOR value
-	unsigned long int data_accum {0};	// Accumulator for received data
+	unsigned long long int data_accum {0};	// Accumulator for received data
 	char16_t calc_crc {0};
 };
